+++ lib/rspec/matchers/built_in/throw_symbol.rb
@@ -102,7 +102,6 @@ module RSpec
       private
 
         def actual_result
-          return "but was not a block" unless Proc === @block
           "got #{caught}"
         end